mcp_client/requirements.txt
fastapi==0.115.12
fastapi-cli==0.0.7
fastmcp==2.5.1
#langchain_core==0.3.62
#langchain_mcp_adapters==0.1.1
#langchain_openai==0.3.18
#langgraph==0.4.7
pydantic==2.11.5
#pymongo==4.13.0
python-dotenv==1.1.0
Requests==2.32.3
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.61
langchain-experimental==0.3.4
langchain-mcp-adapters==0.1.1
langchain-openai==0.3.18
langchain-text-splitters==0.3.8
langgraph==0.4.7
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.1
langgraph-sdk==0.1.70
langsmith==0.3.42
litellm==1.71.2
llama-cloud==0.1.23
llama-cloud-services==0.6.25
llama-index==0.12.39
llama-index-agent-openai==0.4.9
llama-index-cli==0.4.2
llama-index-core==0.12.39
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.7.2
llama-index-llms-ollama==0.6.1
llama-index-llms-openai==0.4.0
llama-index-multi-modal-llms-openai==0.5.1
llama-index-program-openai==0.3.2
llama-index-question-gen-openai==0.3.1
llama-index-readers-file==0.4.8
llama-index-readers-llama-parse==0.4.0
llama-index-tools-mcp==0.2.2
llama-parse==0.6.25
lxml==5.4.0
Markdown==3.8
markdown-it-py==3.0.0
markdown-to-confluence==0.3.4
markdownify==1.1.0
openpyxl==3.1.5